Add `parser.shared.ts`
