+++ packages/shared/src/lib/errorUtils.shared.ts
@@ -1,7 +1,3 @@
-import type {ZodSchema} from 'zod';
-
-import {logger} from '@shared/services/logger.shared';
-
 import type {AsyncResult, ErrorResult, Result} from '@shared/types/result.types';
 import {makeErrorResult, makeSuccessResult} from '@shared/types/result.types';
 import type {Supplier} from '@shared/types/utils.types';
@@ -67,30 +63,6 @@ export function prefixErrorResult(errorResult: ErrorResult, errorPrefix: string)
 export function prefixResultIfError<T>(result: Result<T>, errorPrefix: string): Result<T> {
   if (result.success) return result;
   return makeErrorResult(prefixError(result.error, errorPrefix));
-}
-
-/**
- * Parses a value using a Zod schema and returns a `SuccessResult` with the parsed value if
- * successful, or an `ErrorResult` if the value is invalid.
- */
-export function parseZodResult<T>(zodSchema: ZodSchema<T>, value: unknown): Result<T> {
-  const zodResult = zodSchema.safeParse(value);
-
-  if (!zodResult.success) {
-    logger.error(new Error('Error parsing value with Zod'), {
-      value,
-      error: zodResult.error,
-      issues: zodResult.error.issues,
-    });
-    return makeErrorResult(
-      new Error(
-        `Error parsing value: ${JSON.stringify(zodResult.error.issues.map((issue) => issue.message).join(', '))}`,
-        {cause: zodResult.error}
-      )
-    );
-  }
-
-  return makeSuccessResult(zodResult.data);
 }
 
 /**
