+++ packages/shared/src/parsers/tags.parser.ts
@@ -1,4 +1,5 @@
-import {parseZodResult, prefixErrorResult} from '@shared/lib/errorUtils.shared';
+import {prefixErrorResult} from '@shared/lib/errorUtils.shared';
+import {parseZodResult} from '@shared/lib/parser.shared';
 
 import type {Result} from '@shared/types/result.types';
 import {makeSuccessResult} from '@shared/types/result.types';
