+++ packages/shared/src/parsers/user.parser.ts
@@ -1,6 +1,7 @@
 import type {User as FirebaseUser} from 'firebase/auth';
 
-import {parseZodResult, prefixErrorResult} from '@shared/lib/errorUtils.shared';
+import {prefixErrorResult} from '@shared/lib/errorUtils.shared';
+import {parseZodResult} from '@shared/lib/parser.shared';
 
 import type {Result} from '@shared/types/result.types';
 import {makeErrorResult, makeSuccessResult} from '@shared/types/result.types';
